ml2/ml2/requirements.txt
torch
torchvision
opencv-python
pillow
numpy
